requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.6
Deprecated==1.3.1
Flask==3.1.3
flask-cors==6.0.2
Flask-Limiter==4.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
typing_extensions==4.15.0
Werkzeug==3.1.7
wrapt==2.1.2
razorpay
reportlab
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openpyxl
